Some initial refactoring to bring down complexity
